backend/requirements.txt
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.2.0
langchain-huggingface>=1.2.0
langchain-chroma>=1.1.0
langchain-groq>=1.1.0
langchain-text-splitters>=1.1.0
chromadb>=1.5.0
sentence-transformers>=5.0.0
python-dotenv>=1.0.0
TurkishStemmer>=1.3
unstructured[pdf]>=0.18.0
fastapi>=0.115.0
uvicorn>=0.34.0
